se ha añadido una comprobación a deleteShot, un shot no puede ser borrado si no existe una ocurrencia de él mismo en nuestra base de datos local.  Se ha añadido un método insertShotInTableSync, para la tabla de sincronización

+++ app/src/main/java/gm/mobi/android/db/manager/ShotManager.java
@@ -1,7 +1,9 @@
 package gm.mobi.android.db.manager;
 
 import android.content.ContentValues;
+import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.database.sqlite.SQLiteException;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -9,6 +11,8 @@ import java.util.List;
 import gm.mobi.android.db.GMContract;
 import gm.mobi.android.db.mappers.ShotMapper;
 import gm.mobi.android.db.objects.Shot;
+import gm.mobi.android.db.objects.TableSync;
+import timber.log.Timber;
 
 public class ShotManager {
 
@@ -23,6 +27,7 @@ public class ShotManager {
         }else{
             res = db.insertWithOnConflict(GMContract.ShotTable.TABLE, null, contentValues, SQLiteDatabase.CONFLICT_REPLACE);
         }
+        insertShotInTableSync(db);
     }
 
     /**
@@ -37,7 +42,9 @@ public class ShotManager {
                 res = deleteShot(db, shot);
             }else{
                 res = db.insertWithOnConflict(GMContract.ShotTable.TABLE,null,contentValues,SQLiteDatabase.CONFLICT_REPLACE);
+                Timber.d("Shot inserted", shot.getIdShot());
             }
+            insertShotInTableSync(db);
             db.endTransaction();
         }
    }
@@ -46,6 +53,32 @@ public class ShotManager {
      * Delete a shot
      * */
    public static long deleteShot(SQLiteDatabase db, Shot shot){
-        return db.delete(GMContract.ShotTable.TABLE, GMContract.ShotTable.ID_SHOT,new String[]{String.valueOf(shot.getIdShot())});
+       long res = 0;
+       String args = GMContract.ShotTable.ID_SHOT+"=?";
+       String[] stringArgs = new String[]{String.valueOf(shot.getIdShot())};
+       Cursor c = db.query(GMContract.ShotTable.TABLE, GMContract.ShotTable.PROJECTION, args, stringArgs, null, null, null);
+       if (c.getCount() > 0) {
+           res = db.delete(GMContract.ShotTable.TABLE, GMContract.ShotTable.ID_SHOT, new String[]{String.valueOf(shot.getIdShot())});
+       }
+       c.close();
+       return res;
    }
+
+    public static long insertShotInTableSync(SQLiteDatabase db){
+        TableSync tablesSync = new TableSync();
+        tablesSync.setOrder(3); // It's the third data type the application insert in database
+        tablesSync.setDirection("BOTH");
+        tablesSync.setEntity(GMContract.ShotTable.TABLE);
+        tablesSync.setMax_timestamp(System.currentTimeMillis());
+        if(GeneralManager.isTableEmpty(db, GMContract.ShotTable.TABLE)){
+            tablesSync.setMin_timestamp(System.currentTimeMillis());
+        }
+        //We don't have this information already
+//        tablesSync.setMaxRows();
+//        tablesSync.setMinRows();
+        return SyncTableManager.insertOrUpdateSyncTable(db,tablesSync);
+    }
+
+
+
 }
